fracture: reduce extracted business and user implementations
