+++ frontend/src/lib/components/business/OverviewImpl.css
@@ -1,0 +1,291 @@
+	.overview {
+		max-width: 1400px;
+		margin: 0 auto;
+	}
+
+	.overview-header {
+		margin-bottom: 1.5rem;
+	}
+
+	.overview-header h1 {
+		margin: 0 0 0.25rem 0;
+		font-size: 1.75rem;
+		font-weight: 600;
+	}
+
+	.date-display {
+		color: var(--text-secondary, #888);
+		margin: 0;
+	}
+
+	/* Stats Grid */
+	.stats-grid {
+		display: grid;
+		grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
+		gap: 1rem;
+		margin-bottom: 1.5rem;
+	}
+
+	.stat-card {
+		background: var(--surface-base, #16213e);
+		border-radius: 12px;
+		padding: 1.25rem;
+		cursor: pointer;
+		transition: all 0.2s ease;
+		border: 1px solid var(--border-color, #2a2a4a);
+		position: relative;
+		overflow: hidden;
+	}
+
+	.stat-card:hover {
+		transform: translateY(-2px);
+		border-color: var(--accent, #5865f2);
+	}
+
+	.stat-card.warning .stat-value {
+		color: #f97316;
+	}
+
+	.stat-icon {
+		font-size: 1.5rem;
+		margin-bottom: 0.5rem;
+	}
+
+	.stat-value {
+		font-size: 1.75rem;
+		font-weight: 700;
+		color: var(--text-heading, #fff);
+	}
+
+	.stat-label {
+		font-size: 0.85rem;
+		color: var(--text-secondary, #888);
+		margin-top: 0.25rem;
+	}
+
+	.stat-progress {
+		position: absolute;
+		bottom: 0;
+		left: 0;
+		right: 0;
+		height: 4px;
+		background: var(--surface-raised, #1f2937);
+	}
+
+	.progress-bar {
+		height: 100%;
+		background: var(--accent, #5865f2);
+		transition: width 0.3s ease;
+	}
+
+	/* Content Grid */
+	.content-grid {
+		display: grid;
+		grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
+		gap: 1rem;
+	}
+
+	.content-card {
+		background: var(--surface-base, #16213e);
+		border-radius: 12px;
+		border: 1px solid var(--border-color, #2a2a4a);
+		overflow: hidden;
+	}
+
+	.content-card.wide {
+		grid-column: span 2;
+	}
+
+	.content-card.alert {
+		border-color: #ef4444;
+	}
+
+	.content-card.alert .card-header h2 {
+		color: #ef4444;
+	}
+
+	.card-header {
+		display: flex;
+		justify-content: space-between;
+		align-items: center;
+		padding: 1rem 1.25rem;
+		border-bottom: 1px solid var(--border-color, #2a2a4a);
+	}
+
+	.card-header h2 {
+		margin: 0;
+		font-size: 1rem;
+		font-weight: 600;
+	}
+
+	.view-all {
+		background: transparent;
+		border: none;
+		color: var(--accent, #5865f2);
+		font-size: 0.85rem;
+		cursor: pointer;
+		padding: 0.25rem 0.5rem;
+		border-radius: 4px;
+		transition: background 0.2s;
+	}
+
+	.view-all:hover {
+		background: var(--surface-hover, #1f2937);
+	}
+
+	.badge {
+		background: var(--accent, #5865f2);
+		color: white;
+		font-size: 0.75rem;
+		padding: 0.2rem 0.5rem;
+		border-radius: 12px;
+		font-weight: 600;
+	}
+
+	.card-content {
+		padding: 1rem 1.25rem;
+	}
+
+	.empty-message {
+		color: var(--text-secondary, #888);
+		font-size: 0.9rem;
+		text-align: center;
+		padding: 1rem 0;
+		margin: 0;
+	}
+
+	/* Task List */
+	.task-list, .event-list {
+		list-style: none;
+		padding: 0;
+		margin: 0;
+	}
+
+	.task-item, .event-item {
+		display: flex;
+		align-items: center;
+		gap: 0.75rem;
+		padding: 0.6rem 0;
+		border-bottom: 1px solid var(--border-color, #2a2a4a);
+	}
+
+	.task-item:last-child, .event-item:last-child {
+		border-bottom: none;
+	}
+
+	.priority-dot {
+		width: 8px;
+		height: 8px;
+		border-radius: 50%;
+		flex-shrink: 0;
+	}
+
+	.task-title, .event-title {
+		flex: 1;
+		font-size: 0.9rem;
+		white-space: nowrap;
+		overflow: hidden;
+		text-overflow: ellipsis;
+	}
+
+	.task-time, .task-date {
+		font-size: 0.8rem;
+		color: var(--text-secondary, #888);
+	}
+
+	.task-date.overdue {
+		color: #ef4444;
+	}
+
+	.more-items {
+		text-align: center;
+		color: var(--text-secondary, #888);
+		font-size: 0.85rem;
+		margin: 0.5rem 0 0 0;
+	}
+
+	/* Event List */
+	.event-item {
+		display: flex;
+		gap: 1rem;
+	}
+
+	.event-date-badge {
+		display: flex;
+		flex-direction: column;
+		align-items: center;
+		justify-content: center;
+		width: 48px;
+		height: 48px;
+		border-radius: 8px;
+		color: white;
+		flex-shrink: 0;
+	}
+
+	.event-day {
+		font-size: 1.1rem;
+		font-weight: 700;
+		line-height: 1;
+	}
+
+	.event-month {
+		font-size: 0.65rem;
+		text-transform: uppercase;
+	}
+
+	.event-details {
+		flex: 1;
+		display: flex;
+		flex-direction: column;
+		justify-content: center;
+	}
+
+	.event-title {
+		font-weight: 500;
+	}
+
+	.event-time {
+		font-size: 0.8rem;
+		color: var(--text-secondary, #888);
+	}
+
+	/* Diary Preview */
+	.diary-preview-list {
+		display: grid;
+		grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
+		gap: 1rem;
+	}
+
+	.diary-preview {
+		background: var(--surface-raised, #1f2937);
+		border-radius: 8px;
+		padding: 1rem;
+	}
+
+	.diary-date {
+		font-size: 0.8rem;
+		color: var(--accent, #5865f2);
+		font-weight: 600;
+		margin-bottom: 0.25rem;
+	}
+
+	.mood-badge {
+		font-size: 1.25rem;
+	}
+
+	.diary-excerpt {
+		font-size: 0.85rem;
+		color: var(--text-secondary, #aaa);
+		margin: 0.5rem 0 0 0;
+		line-height: 1.4;
+	}
+
+	@media (max-width: 768px) {
+		.content-card.wide {
+			grid-column: span 1;
+		}
+
+		.stats-grid {
+			grid-template-columns: repeat(2, 1fr);
+		}
+	}
